hub/requirements.txt
google-cloud-pubsub==2.1.0